\name{CallExpmDemo-package}
\alias{CallExpmDemo-package}
\alias{CallExpmDemo}
\docType{package}
\title{
\packageTitle{CallExpmDemo}
}
\description{
\packageDescription{CallExpmDemo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CallExpmDemo}
\packageIndices{CallExpmDemo}

}
\author{

\packageAuthor{CallExpmDemo}

Maintainer: Jonathan Berrisch \url{mailto:Jonathan@Berrisch.biz}

}

\seealso{
\code{\link[expm:expm]{expm}}
}
