% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expm_rcpp}
\alias{expm_rcpp}
\title{Compute the exponential of a matrix}
\usage{
expm_rcpp(x)
}
\arguments{
\item{x}{An numeric matrix}
}
\value{
A numeric matrix
}
\description{
This function computes the exponential of a matrix.
}
\seealso{
The \pkg{expm} package and its documentation.
}
